requirements.txt
# requirements.txt

# Core Telegram libraries
telethon
python-telegram-bot[ext] >= 20.0 # Nên chỉ định phiên bản PTB để tránh lỗi không tương thích sau này

# Environment variables
python-dotenv

# Blockchain interaction
web3 >=6.0.0,<7.0.0 # Phiên bản web3 v6 (cần C++ Build Tools trên Windows)

# Encryption for private keys
cryptography >= 3.4 # Hoặc phiên bản mới hơn

# Dependencies của web3 (thường được cài tự động, nhưng liệt kê nếu cần)
# aiohttp >= 3.7.4 # Dependency của web3 và PTB
# websockets >= 10.0 # Dependency của web3
# eth-account < 0.13, >= 0.8.0 # Dependency của web3
# jsonschema >= 4.0.0 # Dependency của web3
# protobuf >= 4.21.6 # Dependency của web3
# eth-typing < 5.0.0, >= 3.0.0 # Dependency của web3
# eth-utils < 5, >= 2.1.0 # Dependency của web3
# hexbytes < 0.4.0, >= 0.1.0 # Dependency của web3